% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awb.R
\name{awb}
\alias{awb}
\title{Autoregressive wild bootstrap}
\usage{
awb(resid, L)
}
\arguments{
\item{resid}{Residual vector}

\item{L}{Precalculated Cholesky matrix}
}
\value{
Bootstrap errors
}
\description{
Autoregressive wild bootstrap
}
\keyword{internal}
