% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_bootstrap.R
\name{AWBootstrap}
\alias{AWBootstrap}
\title{Autoregressive wild bootstrap for nonparametric trend}
\usage{
AWBootstrap(y, time, m.tilde, B, h, newtime, k)
}
\description{
Autoregressive wild bootstrap for nonparametric trend
}
\keyword{internal}
