% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_detection.R
\name{break.teststat}
\alias{break.teststat}
\title{Break test statistic}
\usage{
break.teststat(y, time, nterms, trim.frac)
}
\arguments{
\item{y}{Numeric vector of observations.}

\item{time}{Numeric vector of time indices.}

\item{nterms}{Integer specifying the number of Fourier terms.}

\item{trim.frac}{Trimming fraction for break estimation.}
}
\description{
Break test statistic
}
