% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholesky.R
\name{cholesky.decomp}
\alias{cholesky.decomp}
\title{Cholesky decomposition for dependent bootstrap errors}
\usage{
cholesky.decomp(time, theta, l)
}
\arguments{
\item{time}{Time index}

\item{theta}{Dependence parameter}

\item{l}{Scaling parameter}
}
\value{
Cholesky factor
}
\description{
Cholesky decomposition for dependent bootstrap errors
}
\keyword{internal}
