% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k.cosine}
\alias{k.cosine}
\title{Cosine kernel}
\usage{
k.cosine(u)
}
\arguments{
\item{u}{Numeric vector}
}
\value{
Kernel weights
}
\description{
Usage note: internal function, if using cosine kernel run
k <- npTrend::k.cosine and pass k into functions which
use k as input
}
