% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k.epanech}
\alias{k.epanech}
\title{Epanechnikov kernel}
\usage{
k.epanech(u)
}
\arguments{
\item{u}{Numeric vector}
}
\value{
Kernel weights
}
\description{
Usage note: internal function, if using Epanechnikov kernel
run k <- npTrend::k.epanech and pass k into functions which
use k as input
}
