% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k.quartic}
\alias{k.quartic}
\title{Quartic kernel}
\usage{
k.quartic(u)
}
\arguments{
\item{u}{Numeric vector}
}
\value{
Kernel weights
}
\description{
Usage note: internal function, if using quartic kernel run
k <- npTrend::k.quartic and pass k into functions which
use k as input
}
