% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k.unif}
\alias{k.unif}
\title{Uniform kernel}
\usage{
k.unif(u)
}
\arguments{
\item{u}{Numeric vector}
}
\value{
Kernel weights
}
\description{
Usage note: internal function, if using uniform kernel run
k <- npTrend::k.unif and pass k into functions which
use k as input
}
