% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_trend.R
\name{lin.trend}
\alias{lin.trend}
\title{Main linear trend estimation with break testing}
\usage{
lin.trend(y, time, B, alpha, nterms = 3, trim.frac = 0.1, cores = 1)
}
\arguments{
\item{y}{Numeric vector of length n of observations containing the time series of interest}

\item{time}{A vector of length n, containing the corresponding dates in decimal format}

\item{B}{Integer number of bootstrap replications.}

\item{alpha}{Nominal coverage of confidence intervals will be (1-alpha)}

\item{nterms}{Integer specifying the number of Fourier terms, default = 3}

\item{trim.frac}{Parameter pi as described in Section 3.1 of Friedrich et al. (2020), default equals 0.1}

\item{cores}{Integer; number of CPU cores to use for parallel bootstrap. Default is 1.}
}
\value{
A list containing the following:
\item{p}{Numeric. p-value of bootstrap test for break in trend.}
\item{cv}{Numeric. Critical value of the test.}
\item{S_T}{Numeric. Test statistic S_T.}
\item{para}{Numeric vector. Parameters of the trend.}
\item{fit}{List. Fitted values.}
\item{breakpoint}{Numeric. Point estimation of trend break location.}
\item{CI.break}{Numeric vector. Bootstrap 95\% confidence interval for trend break.}
\item{CI.para}{List. Bootstrap 95\% confidence intervals for parameters.}
}
\description{
Trend estimation with bootstrap inference for location and confidence interval of the break, and point estimation and confidence intervals for parameters
}
