% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_estimators.R
\name{LLEstimation}
\alias{LLEstimation}
\title{Optional: Local linear nonparametric estimator}
\usage{
LLEstimation(y, time, h, k)
}
\description{
Optional: Local linear nonparametric estimator
}
\keyword{internal}
