% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_trend.R
\name{nonpara.trend}
\alias{nonpara.trend}
\title{Nonparametric trend estimation with bootstrap inference}
\usage{
nonpara.trend(
  y,
  time,
  B,
  alpha,
  h,
  k,
  nterms = 3,
  G = seq(1, length(y), 1),
  C = 0.5,
  l = 5,
  grid = seq(0.01, by = 0.005, length.out = 50),
  cores = 1
)
}
\arguments{
\item{y}{Numeric vector of length n of observations containing the time series of interest}

\item{time}{A vector of length n, containing the corresponding dates in decimal format}

\item{B}{Integer number of bootstrap replications.}

\item{alpha}{Nominal coverage of confidence intervals will be (1-alpha)}

\item{h}{Bandwidth parameter for nonparametric estimation, can be selected by the user or by a data-driven selection method,
called Modified Cross Validation (MCV) and will be started by selecting "h=-1" with parameter
l=4 and a grid [0.001,0.099] by default. Both can be changed by the user}

\item{k}{Kernel function for nonparametric estimation, with Epanechnikov, Uniform, Quartic, Cosine kernels available}

\item{nterms}{Integer specifying the number of Fourier terms, default = 3}

\item{G}{Set of time indices over which the confidence bands will be made simultaneous, with default being whole sample}

\item{C}{Parameter determining the extend of oversmoothing in Step 1 of the bootstrap algorithm, default is 0.5}

\item{l}{Parameter for MCV, default equals 4}

\item{grid}{Range of possible bandwidth for MCV, default being [0.001,0.099] in steps of 0.002}

\item{cores}{cores Integer; number of CPU cores to use for parallel bootstrap. Default is 1.}
}
\value{
A list of class "results.nonpara" containing:
\item{m.hat}{Trend estimate vector.}
\item{confidence.pw}{Matrix of pointwise confidence intervals.}
\item{confidence.simu}{Matrix of simultaneous confidence bands.}
\item{h.opt}{Selected bandwidth.}
\item{CV}{MCV criterion matrix.}
}
\description{
Estimates trend, with bootstrap pointwise confidence intervals and simultaneous confidence bands for the trend
}
