% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_intervals.R
\name{pointwise.int}
\alias{pointwise.int}
\title{Pointwise confidence intervals}
\usage{
pointwise.int(y, m.hat, m.star, m.tilde, alpha, B)
}
\description{
Pointwise confidence intervals
}
\keyword{internal}
