% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressors.R
\name{regressors}
\alias{regressors}
\title{Construct regressor matrix for trend with break and Fourier terms}
\usage{
regressors(breakdate, time, nterms)
}
\description{
Construct regressor matrix for trend with break and Fourier terms
}
\keyword{internal}
