% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_fourier.R
\name{remove.fourier}
\alias{remove.fourier}
\title{Regress data on Fourier terms and obtain residuals}
\usage{
remove.fourier(y, time, nterms)
}
\description{
Regress data on Fourier terms and obtain residuals
}
\keyword{internal}
