% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_intervals.R
\name{simult.int}
\alias{simult.int}
\title{Simultaneous confidence bands}
\usage{
simult.int(y, m.hat, m.tilde, m.star, alpha, B, G, G.total)
}
\description{
Simultaneous confidence bands
}
\keyword{internal}
