% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_trend.R
\name{plot.results.linear}
\alias{plot.results.linear}
\title{Plot linear trend results}
\usage{
\method{plot}{results.linear}(x, time, fit, breakdate, CI.break, low, up, xlabel, ylabel, ...)
}
\arguments{
\item{x}{An object of class "results.linear".}

\item{time}{A vector containing the corresponding dates.}

\item{fit}{Fitted values (optional, defaults to x[[5]]).}

\item{breakdate}{Estimated break date (optional).}

\item{CI.break}{Confidence interval for break (optional).}

\item{low}{Lower bound of y-axis.}

\item{up}{Upper bound of y-axis.}

\item{xlabel}{X-axis label.}

\item{ylabel}{Y-axis label.}

\item{...}{Additional arguments passed to the generic \code{plot} function.}
}
\value{
Returns plots of the linear trend estimation, estimated breakpoint, and confidence intervals
}
\description{
Plot linear trend results
}
