% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_trend.R
\name{plot.results.nonpara}
\alias{plot.results.nonpara}
\title{Plot nonparametric trend and confidence bands}
\usage{
\method{plot}{results.nonpara}(x, time, m.hat, int.s, low, up, xlabel, ylabel, ...)
}
\arguments{
\item{x}{An object of class "results.linear".}

\item{time}{A vector containing the corresponding dates.}

\item{m.hat}{Trend estimate vector.}

\item{int.s}{Matrix of simultaneous confidence bands.}

\item{low}{Lower bound of y-axis.}

\item{up}{Upper bound of y-axis.}

\item{xlabel}{X-axis label.}

\item{ylabel}{Y-axis label.}

\item{...}{Additional arguments passed to the generic \code{plot} function.}
}
\description{
Plot nonparametric trend and confidence bands
}
